requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mega.py==1.0.8
pathlib==1.0.1
pycryptodome==3.23.0
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
tenacity==9.1.2
urllib3==2.4.0
Werkzeug==3.1.3